refactor: seperate service from controller
